// src/types/ResponseTypes/SearchedAlbumsResponse.ts
export interface SearchedAlbumsResponse {
  albums: Albums;
}

export interface Albums {
  href: string;
  items: Item[];
  limit: number;
  next: string;
  offset: number;
  previous: null;
  total: number;
}

export interface Item {
  album_type: AlbumTypeEnum;
  artists: Artist[];
  available_markets: string[];
  external_urls: ExternalUrls;
  href: string;
  id: string;
  images: Image[];
  name: string;
  release_date: string;
  release_date_precision: ReleaseDatePrecision;
  total_tracks: number;
  type: AlbumTypeEnum;
  uri: string;
}

export type AlbumTypeEnum = "album" | "compilation" | "single";

export interface Artist {
  external_urls: ExternalUrls;
  href: string;
  id: string;
  name: string;
  type: ArtistType;
  uri: string;
}

export interface ExternalUrls {
  spotify: string;
}

export type ArtistType = "artist";

export interface Image {
  height: number;
  url: string;
  width: number;
}

export type ReleaseDatePrecision = "day" | "year" | "month";